requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
imaplib2
email-validator
python-dotenv
schedule
flask
flask-cors
beautifulsoup4
requests
numpy
pandas
tiktoken